MessagePipeTestBase refactor
