+++ udmis/src/test/java/com/google/bos/udmi/service/messaging/impl/LocalMessagePipeTest.java
@@ -16,7 +16,7 @@ import udmi.schema.EndpointConfiguration.Protocol;
 /**
  * Tests for LocalMessagePipe.
  */
-public class LocalMessagePipeTest extends MessageTestBase {
+public class LocalMessagePipeTest extends MessagePipeTestBase {
 
   private Map<String, Object> testSend(Object message) {
     getTestDispatcher().publish(message);
