+++ udmis/src/test/java/com/google/bos/udmi/service/messaging/impl/MessageDispatcherImplTest.java
@@ -1,8 +1,8 @@
 package com.google.bos.udmi.service.messaging.impl;
 
-import static com.google.bos.udmi.service.messaging.impl.MessageTestBase.TEST_DESTINATION;
-import static com.google.bos.udmi.service.messaging.impl.MessageTestBase.TEST_NAMESPACE;
-import static com.google.bos.udmi.service.messaging.impl.MessageTestBase.TEST_SOURCE;
+import static com.google.bos.udmi.service.messaging.impl.MessagePipeTestBase.TEST_DESTINATION;
+import static com.google.bos.udmi.service.messaging.impl.MessagePipeTestBase.TEST_NAMESPACE;
+import static com.google.bos.udmi.service.messaging.impl.MessagePipeTestBase.TEST_SOURCE;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
